add SOLID requirements
